+++ tests/test_state_managed_abc.py
@@ -62,6 +62,7 @@ class TestStateManagedABCs:
             "target-branch",
             "000000000",
             "111111111",
+            "222222222",
             Member("name", "email", "123"),
             datetime.now(),
             Repo("123", "name", default_branch="target-branch"),